refactor: extract in-memory tests
