+++ test/in-memory/prevalence-in-memory.test.ts
@@ -1,13 +1,13 @@
 import { assertEquals } from "https://deno.land/std@0.203.0/assert/assert_equals.ts";
 import { assertInstanceOf } from "https://deno.land/std@0.203.0/assert/assert_instance_of.ts";
 import { describe } from "https://deno.land/std@0.203.0/testing/bdd.ts";
-import { Action } from "../mod.ts";
-import { AddPostAction } from "./add-post-action.ts";
-import { AddUserAction } from "./add-user-action.ts";
-import { MyModel, Post, User } from "./fixture-types.ts";
+import { Action } from "../../mod.ts";
+import { AddPostAction } from "../add-post-action.ts";
+import { AddUserAction } from "../add-user-action.ts";
+import { MyModel, Post, User } from "../fixture-types.ts";
 import { test } from "./test-case.ts";
 
-describe("prevalence", () => {
+describe("prevalence in :memory:", () => {
   test((system) => assertEquals(system.name, "test"));
   test((system) => assertEquals(system.model.posts.length, 2));
   test((system) => assertEquals(system.model.users.length, 26));
